requirements.txt
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==9.5.0
Pillow==10.0.0
torch==2.1.0.dev20230622+cu118
torchvision==0.16.0.dev20230622+cu118
